Rearrange and rename PyVOLCANS test files
